requirements.txt
Django>=4.2.0
requests>=2.28.0
amadeus==7.1.0
python-dotenv>=0.19.0
djangorestframework>=3.14.0
Pillow>=9.0.0